requirements.txt
Django==3.2.8
djangorestframework==3.12.4
instaloader==4.9
requests==2.26.0